feat: create station search strategy for genre
