+++ frontend/src/App.tsx
@@ -1,24 +1,23 @@
 import "./App.css"
 import { useRef, useState } from "react"
 import { toast, Toaster } from "sonner"
-import { getRandomStation } from "./api/radiobrowser/station"
 import Map from "./features/map/components/Map/Map"
 import RadioSelect from "./features/radioselect/components/RadioSelect/RadioSelect"
 import Header from "./components/Header/Header"
 import Footer from "./components/Footer/Footer"
 import { Station } from "./api/radiobrowser/types"
-import { GenreInformation } from "./api/radiobrowser/genreTags"
+import { StationSearchStrategy } from "./api/radiobrowser/searchStrategy/StationSearchStrategy"
 
 function App() {
   const abortControllerRef = useRef<AbortController | null>(null)
   const [currentStation, setCurrentStation] = useState<Station | null>(null)
   const [isLoading, setIsLoading] = useState<boolean>(false)
 
-  async function displayRandomStation(genre: GenreInformation) {
+  async function displayRandomStation(searchStrategy: StationSearchStrategy) {
     setIsLoading(true)
     abortControllerRef.current?.abort()
     abortControllerRef.current = new AbortController()
-    const station = await getRandomStation(abortControllerRef.current, genre)
+    const station = await searchStrategy.findStation(abortControllerRef.current)
     if (station) {
       setCurrentStation(station)
     } else {
