+++ frontend/src/api/radiobrowser/searchStrategy/StationSearchStrategy.ts
@@ -1,0 +1,5 @@
+import { Station } from "../types"
+
+export interface StationSearchStrategy {
+  findStation: (abortController: AbortController) => Promise<Station | null>
+}
