+++ frontend/src/api/radiobrowser/station.ts
@@ -1,21 +1,12 @@
 import ky from "ky"
-import { getRandomServer } from "./servers"
 import { Station } from "./types"
 import { convertMissingInformation } from "./converter/stationConverter"
-import { GenreInformation } from "./genreTags"
 
-async function getRandomStation(
+async function getStation(
   abortController: AbortController,
-  genre: GenreInformation
+  url: string,
+  searchParams: URLSearchParams
 ): Promise<Station | null> {
-  const tag = genre.searchTag
-  const offset = getRandomInt(genre.approxStationCount)
-  const limit = 3
-  const searchParams = new URLSearchParams(
-    `?order=random&limit=${limit}&hidebroken=true&is_https=true&offset=${offset}&tag=${tag}`
-  )
-  const server = await getRandomServer()
-  const url = `${server}/json/stations/search`
   try {
     const json: Station[] = await ky
       .get(url, { retry: 0, signal: abortController.signal, searchParams })
@@ -37,4 +28,4 @@ function getRandomInt(max: number) {
   return Math.floor(Math.random() * max)
 }
 
-export { getRandomStation }
+export { getStation }
